WIP: PR Review Refactor
