+++ pkg/migrate/migrate_test.go
@@ -3,7 +3,6 @@ package migrate
 import (
 	"context"
 	"fmt"
-	"io"
 	"log"
 	"strings"
 	"testing"
@@ -18,7 +17,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	k8sclient "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/fake"
-	"k8s.io/client-go/rest"
 	"k8s.io/utils/pointer"
 )
 
@@ -922,34 +920,6 @@ func TestGetPVCs(t *testing.T) {
 			req.Equal(test.originalPVCs, originalPVCs)
 			req.Equal(test.namespaces, nses)
 		})
-	}
-}
-
-func TestNewPVMigrator(t *testing.T) {
-	// test empty logger
-	_, err := NewPVMigrator(&rest.Config{}, nil, "src", "dst", 0)
-	if err == nil || err.Error() != "no logger provided" {
-		t.Errorf("expected failure creating object: %v", err)
-	}
-
-	logger := log.New(io.Discard, "", 0)
-
-	// test src storage class
-	_, err = NewPVMigrator(&rest.Config{}, logger, "", "dst", 0)
-	if err == nil || err.Error() != "empty source storage class" {
-		t.Errorf("expected failure creating object: %v", err)
-	}
-
-	// test empty dst sc
-	_, err = NewPVMigrator(&rest.Config{}, logger, "src", "", 0)
-	if err == nil || err.Error() != "empty destination storage class" {
-		t.Errorf("expected failure creating object: %v", err)
-	}
-
-	// happy path
-	_, err = NewPVMigrator(&rest.Config{}, logger, "src", "dst", 0)
-	if err != nil {
-		t.Errorf("unexpected failure creating object: %v", err)
 	}
 }
 
@@ -3092,769 +3062,6 @@ func Test_copyAllPVCs(t *testing.T) {
 				return
 			}
 			req.NoError(err)
-		})
-	}
-}
-
-func Test_validateVolumeAccessModes(t *testing.T) {
-	for _, tt := range []struct {
-		name            string
-		srcStorageClass string
-		dstStorageClass string
-		deletePVTimeout time.Duration
-		podReadyTimeout time.Duration
-		wantErr         bool
-		resources       []runtime.Object
-		input           map[string]corev1.PersistentVolume
-		expected        map[string]map[string]PVCError
-	}{
-		{
-			name: "With compatible access modes, expect no errors",
-			input: map[string]corev1.PersistentVolume{
-				"pv0": {
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pv0",
-					},
-					Spec: corev1.PersistentVolumeSpec{
-						StorageClassName: "srcSc",
-						ClaimRef: &corev1.ObjectReference{
-							Name: "pvc",
-						},
-					},
-				},
-			},
-			expected:        make(map[string]map[string]PVCError),
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources: []runtime.Object{
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "srcSc",
-					},
-				},
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "dstSc",
-					},
-				},
-				&corev1.PersistentVolumeClaim{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pvc",
-					},
-				},
-				&corev1.PersistentVolume{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pv0",
-					},
-					Spec: corev1.PersistentVolumeSpec{
-						StorageClassName: "srcSc",
-						ClaimRef: &corev1.ObjectReference{
-							Name: "pvc",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "When there is no matching pvc for a pv, expect error",
-			input: map[string]corev1.PersistentVolume{
-				"pv0": {
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pv0",
-					},
-					Spec: corev1.PersistentVolumeSpec{
-						StorageClassName: "srcSc",
-						ClaimRef: &corev1.ObjectReference{ // this pvc will not be found
-							Name:      "pvc",
-							Namespace: "default",
-						},
-					},
-				},
-			},
-			expected:        nil,
-			wantErr:         true,
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources: []runtime.Object{
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "srcSc",
-					},
-				},
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "dstSc",
-					},
-					Provisioner:       "kubernetes.io/no-provisioner",
-					VolumeBindingMode: (*storagev1.VolumeBindingMode)(pointer.String("WaitForFirstConsumer")),
-				},
-				&corev1.PersistentVolumeClaim{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "pvc",
-						Namespace: "ns1",
-					},
-					Spec: corev1.PersistentVolumeClaimSpec{
-						StorageClassName: pointer.String("srcSc"),
-						AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteAll"},
-						Resources: corev1.ResourceRequirements{
-							Requests: corev1.ResourceList{
-								corev1.ResourceStorage: resource.MustParse("1Mi"),
-							},
-						},
-					},
-				},
-				&corev1.PersistentVolume{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pv0",
-					},
-					Spec: corev1.PersistentVolumeSpec{
-						StorageClassName: "srcSc",
-						ClaimRef: &corev1.ObjectReference{
-							Name:      "pvc",
-							Namespace: "ns1",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "When destination storage class is not found, expect error",
-			input: map[string]corev1.PersistentVolume{
-				"pv0": {
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pv0",
-					},
-					Spec: corev1.PersistentVolumeSpec{
-						StorageClassName: "srcSc",
-						ClaimRef: &corev1.ObjectReference{
-							Name:      "pvc",
-							Namespace: "default",
-						},
-					},
-				},
-			},
-			expected:        nil,
-			wantErr:         true,
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources: []runtime.Object{
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "srcSc",
-					},
-				},
-			},
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			req := require.New(t)
-			kcli := fake.NewSimpleClientset(tt.resources...)
-			testlog := log.New(testWriter{t: t}, "", 0)
-			pvm := PVMigrator{
-				ctx:             context.Background(),
-				log:             testlog,
-				k8scli:          kcli,
-				srcSc:           tt.srcStorageClass,
-				dstSc:           tt.dstStorageClass,
-				deletePVTimeout: 1 * time.Millisecond,
-				podReadyTimeout: 1 * time.Millisecond,
-			}
-			result, err := pvm.ValidateVolumeAccessModes(tt.input)
-			if err != nil {
-				if tt.wantErr {
-					req.Error(err)
-				} else {
-					req.NoError(err)
-				}
-			}
-			req.Equal(result, tt.expected)
-		})
-	}
-}
-
-func Test_getPvcError(t *testing.T) {
-	for _, tt := range []struct {
-		name            string
-		srcStorageClass string
-		dstStorageClass string
-		deletePVTimeout time.Duration
-		podTimeout      time.Duration
-		wantErr         bool
-		resources       []runtime.Object
-		input           *corev1.PersistentVolumeClaim
-		expected        PVCError
-	}{
-		{
-			name: "When there is a PVC error, expect ProvisioningFailed event",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "pvc",
-					Namespace: "default",
-					UID:       "12345",
-				},
-				Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimPending},
-			},
-			expected: PVCError{
-				reason:  "ProvisioningFailed",
-				from:    "kubernetes.io/no-provisioner",
-				message: "Only support ReadWriteOnce access mode",
-			},
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources: []runtime.Object{
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "srcSc",
-					},
-				},
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "dstSc",
-					},
-				},
-				&corev1.PersistentVolumeClaim{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "pvc",
-						Namespace: "default",
-						UID:       "12345",
-					},
-					Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimPending},
-				},
-				&corev1.EventList{
-					Items: []corev1.Event{
-						{
-							ObjectMeta: metav1.ObjectMeta{
-								Name:      "pvc-error-event",
-								Namespace: "default",
-							},
-							InvolvedObject: corev1.ObjectReference{
-								Kind:      "PersistentVolumeClaim",
-								Namespace: "default",
-								Name:      "pvc",
-								UID:       "12345",
-							},
-							Source:  corev1.EventSource{Component: "kubernetes.io/no-provisioner"},
-							Reason:  "ProvisioningFailed",
-							Message: "Only support ReadWriteOnce access mode",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "When PVC event reason is not ProvisioningFailed, expect error",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "pvc",
-					Namespace: "default",
-					UID:       "12345",
-				},
-				Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimPending},
-			},
-			expected:        PVCError{},
-			wantErr:         true,
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources: []runtime.Object{
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "srcSc",
-					},
-				},
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "dstSc",
-					},
-				},
-				&corev1.PersistentVolumeClaim{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "pvc",
-						Namespace: "default",
-						UID:       "12345",
-					},
-					Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimPending},
-				},
-				&corev1.EventList{
-					Items: []corev1.Event{
-						{
-							ObjectMeta: metav1.ObjectMeta{
-								Name:      "pvc-error-event",
-								Namespace: "default",
-							},
-							InvolvedObject: corev1.ObjectReference{
-								Kind:      "PersistentVolumeClaim",
-								Namespace: "default",
-								Name:      "pvc",
-								UID:       "12345",
-							},
-							Source:  corev1.EventSource{Component: "kubernetes.io/no-provisioner"},
-							Reason:  "Provisioning",
-							Message: "External provisioner is provisiong volume for claim pvc",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "When PVC is pending due to an error but there are no events for it, expect error",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "pvc",
-					Namespace: "default",
-					UID:       "12345",
-				},
-				Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimPending},
-			},
-			expected:        PVCError{},
-			wantErr:         true,
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources: []runtime.Object{
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "srcSc",
-					},
-				},
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "dstSc",
-					},
-				},
-				&corev1.PersistentVolumeClaim{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "pvc",
-						Namespace: "default",
-						UID:       "12345",
-					},
-					Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimPending},
-				},
-			},
-		},
-		{
-			name: "When PVC is not in Pending status, expect error",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "pvc",
-					Namespace: "default",
-					UID:       "12345",
-				},
-				Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimBound},
-			},
-			expected:        PVCError{},
-			wantErr:         true,
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources:       []runtime.Object{},
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			req := require.New(t)
-			kcli := fake.NewSimpleClientset(tt.resources...)
-			testlog := log.New(testWriter{t: t}, "", 0)
-			pvm := PVMigrator{
-				ctx:             context.Background(),
-				log:             testlog,
-				k8scli:          kcli,
-				srcSc:           tt.srcStorageClass,
-				dstSc:           tt.dstStorageClass,
-				deletePVTimeout: 1 * time.Millisecond,
-				podReadyTimeout: 1 * time.Millisecond,
-			}
-			result, err := pvm.getPvcError(tt.input)
-			if err != nil {
-				if tt.wantErr {
-					req.Error(err)
-				} else {
-					req.NoError(err)
-				}
-			}
-			req.Equal(tt.expected, result)
-		})
-	}
-}
-
-func Test_checkVolumeAccessModes(t *testing.T) {
-	for _, tt := range []struct {
-		name            string
-		srcStorageClass string
-		dstStorageClass string
-		deletePVTimeout time.Duration
-		podTimeout      time.Duration
-		wantErr         bool
-		resources       []runtime.Object
-		input           *corev1.PersistentVolumeClaim
-		expected        PVCError
-	}{
-		{
-			name: "When the PVC access mode is not supported by destination storage provider, expect PVCError",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "pvc",
-					Namespace: "default",
-					UID:       "12345",
-				},
-				Spec: corev1.PersistentVolumeClaimSpec{
-					StorageClassName: pointer.String("srcSc"),
-					AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteMany"},
-				},
-				Status: corev1.PersistentVolumeClaimStatus{Phase: corev1.ClaimPending},
-			},
-			expected: PVCError{
-				reason:  "ProvisioningFailed",
-				from:    "kubernetes.io/no-provisioner",
-				message: "Only support ReadWriteOnce access mode",
-			},
-			srcStorageClass: "srcSc",
-			dstStorageClass: "dstSc",
-			resources: []runtime.Object{
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "srcSc",
-					},
-				},
-				&storagev1.StorageClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "dstSc",
-					},
-				},
-				&corev1.EventList{
-					Items: []corev1.Event{
-						{
-							ObjectMeta: metav1.ObjectMeta{
-								Name:      "pvc-error-event",
-								Namespace: "default",
-							},
-							InvolvedObject: corev1.ObjectReference{
-								Kind:      "PersistentVolumeClaim",
-								Namespace: "default",
-								UID:       "12345",
-							},
-							Source:  corev1.EventSource{Component: "kubernetes.io/no-provisioner"},
-							Reason:  "ProvisioningFailed",
-							Message: "Only support ReadWriteOnce access mode",
-						},
-					},
-				},
-				&corev1.Pod{
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "Pod",
-						APIVersion: "v1",
-					},
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "tmpPod",
-						Namespace: "default",
-					},
-					Spec: corev1.PodSpec{
-						RestartPolicy: corev1.RestartPolicyNever,
-						Containers: []corev1.Container{
-							{
-								Name:  "busybox",
-								Image: "busybox",
-								Command: []string{
-									"sleep",
-									"3600",
-								},
-							},
-						},
-					},
-					Status: corev1.PodStatus{Phase: corev1.PodPending},
-				},
-			},
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			req := require.New(t)
-			kcli := fake.NewSimpleClientset(tt.resources...)
-			testlog := log.New(testWriter{t: t}, "", 0)
-			pvm := PVMigrator{
-				ctx:             context.Background(),
-				log:             testlog,
-				k8scli:          kcli,
-				srcSc:           tt.srcStorageClass,
-				dstSc:           tt.dstStorageClass,
-				deletePVTimeout: 1 * time.Millisecond,
-				podReadyTimeout: 1 * time.Millisecond,
-				podNameOverride: "tmpPod",
-			}
-			result, err := pvm.checkVolumeAccessModes(*tt.input)
-			if err != nil {
-				if tt.wantErr {
-					req.Error(err)
-				} else {
-					req.NoError(err)
-				}
-			}
-			req.Equal(tt.expected, result)
-		})
-	}
-}
-
-func Test_buildTmpPVC(t *testing.T) {
-	for _, tt := range []struct {
-		name            string
-		pvcNameOverride string
-		dstStorageClass string
-		input           *corev1.PersistentVolumeClaim
-		expectedPVC     *corev1.PersistentVolumeClaim
-		expectedName    string
-	}{
-		{
-			name: "When PVC name is not overridden, expect unique temp pvc name",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pvc",
-					Namespace: "default",
-				},
-				Spec: corev1.PersistentVolumeClaimSpec{
-					StorageClassName: pointer.String("dstSc"),
-					AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteOnce"},
-				},
-			},
-			expectedPVC: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "default",
-				},
-				Spec: corev1.PersistentVolumeClaimSpec{
-					StorageClassName: pointer.String("dstSc"),
-					AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteOnce"},
-					Resources: corev1.ResourceRequirements{
-						Requests: corev1.ResourceList{
-							corev1.ResourceStorage: resource.MustParse("1Mi"),
-						},
-					},
-				},
-			},
-			expectedName:    "pvmigrate-claim-test-pvc-",
-			dstStorageClass: "dstSc",
-		},
-		{
-			name: "When PVC name is longer than 63 chars, expect name to be trimmed",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "really-long-pvc-name-that-should-be-trimmed-to-avoid-an-error",
-					Namespace: "default",
-				},
-				Spec: corev1.PersistentVolumeClaimSpec{
-					StorageClassName: pointer.String("dstSc"),
-					AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteOnce"},
-				},
-			},
-			expectedPVC: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "default",
-				},
-				Spec: corev1.PersistentVolumeClaimSpec{
-					StorageClassName: pointer.String("dstSc"),
-					AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteOnce"},
-					Resources: corev1.ResourceRequirements{
-						Requests: corev1.ResourceList{
-							corev1.ResourceStorage: resource.MustParse("1Mi"),
-						},
-					},
-				},
-			},
-			expectedName:    "pvmigrate-claim-really-long-pvc-trimmed-to-avoid-an-error-",
-			dstStorageClass: "dstSc",
-		},
-		{
-			name: "When PVC name is overriden, expect non-UID generated name",
-			input: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "my-test-pvc",
-					Namespace: "default",
-				},
-				Spec: corev1.PersistentVolumeClaimSpec{
-					StorageClassName: pointer.String("dstSc"),
-					AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteOnce"},
-				},
-			},
-			expectedPVC: &corev1.PersistentVolumeClaim{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "default",
-				},
-				Spec: corev1.PersistentVolumeClaimSpec{
-					StorageClassName: pointer.String("dstSc"),
-					AccessModes:      []corev1.PersistentVolumeAccessMode{"ReadWriteOnce"},
-					Resources: corev1.ResourceRequirements{
-						Requests: corev1.ResourceList{
-							corev1.ResourceStorage: resource.MustParse("1Mi"),
-						},
-					},
-				},
-			},
-			pvcNameOverride: "pvc-name-override",
-			expectedName:    "pvc-name-override",
-			dstStorageClass: "dstSc",
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			req := require.New(t)
-			var pvcNameOverride string
-			if tt.pvcNameOverride != "" {
-				pvcNameOverride = tt.pvcNameOverride
-			}
-			pvm := PVMigrator{
-				dstSc:           tt.dstStorageClass,
-				pvcNameOverride: pvcNameOverride,
-			}
-			pvc := pvm.buildTmpPVC(*tt.input, tt.dstStorageClass)
-			req.True(strings.HasPrefix(pvc.Name, tt.expectedName))
-			req.Equal(tt.expectedPVC.Spec, pvc.Spec)
-		})
-	}
-}
-
-func Test_buildPVConsumerPod(t *testing.T) {
-	for _, tt := range []struct {
-		name            string
-		podNameOverride string
-		pvcName         string
-		expectedPod     *corev1.Pod
-		expectedName    string
-	}{
-		{
-			name:    "When pod name not overriden, expect unique pod name",
-			pvcName: "test-pvc",
-			expectedPod: &corev1.Pod{
-				TypeMeta: metav1.TypeMeta{
-					Kind:       "Pod",
-					APIVersion: "v1",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					RestartPolicy: corev1.RestartPolicyNever,
-					Volumes: []corev1.Volume{
-						{
-							Name: "tmp",
-							VolumeSource: corev1.VolumeSource{
-								PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-									ClaimName: "test-pvc",
-								},
-							},
-						},
-					},
-					Containers: []corev1.Container{
-						{
-							Name:  "busybox",
-							Image: "busybox",
-							Command: []string{
-								"sleep",
-								"3600",
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									MountPath: "/tmpmount",
-									Name:      "tmp",
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedName: "pvmigrate-vol-consumer-test-pvc-",
-		},
-		{
-			name:    "When pod name is longer than 63 chars, expect pod name to be trimmed",
-			pvcName: "pvc-name-that-should-be-trimmed-because-it-will-cause-an-err",
-			expectedPod: &corev1.Pod{
-				TypeMeta: metav1.TypeMeta{
-					Kind:       "Pod",
-					APIVersion: "v1",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					RestartPolicy: corev1.RestartPolicyNever,
-					Volumes: []corev1.Volume{
-						{
-							Name: "tmp",
-							VolumeSource: corev1.VolumeSource{
-								PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-									ClaimName: "pvc-name-that-should-be-trimmed-because-it-will-cause-an-err",
-								},
-							},
-						},
-					},
-					Containers: []corev1.Container{
-						{
-							Name:  "busybox",
-							Image: "busybox",
-							Command: []string{
-								"sleep",
-								"3600",
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									MountPath: "/tmpmount",
-									Name:      "tmp",
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedName: "pvmigrate-vol-consumer-pvc-namecause-it-will-cause-an-err-",
-		},
-		{
-			name:    "When pod name is overriden, expect non-UID name",
-			pvcName: "test-pvc",
-			expectedPod: &corev1.Pod{
-				TypeMeta: metav1.TypeMeta{
-					Kind:       "Pod",
-					APIVersion: "v1",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					RestartPolicy: corev1.RestartPolicyNever,
-					Volumes: []corev1.Volume{
-						{
-							Name: "tmp",
-							VolumeSource: corev1.VolumeSource{
-								PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-									ClaimName: "test-pvc",
-								},
-							},
-						},
-					},
-					Containers: []corev1.Container{
-						{
-							Name:  "busybox",
-							Image: "busybox",
-							Command: []string{
-								"sleep",
-								"3600",
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									MountPath: "/tmpmount",
-									Name:      "tmp",
-								},
-							},
-						},
-					},
-				},
-			},
-			podNameOverride: "my pod name override",
-			expectedName:    "my pod name override",
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			req := require.New(t)
-			var podNameOverride string
-			if tt.podNameOverride != "" {
-				podNameOverride = tt.podNameOverride
-			}
-			pvm := PVMigrator{
-				podNameOverride: podNameOverride,
-			}
-			pod := pvm.buildPVConsumerPod(tt.pvcName)
-			req.True(strings.HasPrefix(pod.Name, tt.expectedName))
-			req.Equal(tt.expectedPod.Spec, pod.Spec)
 		})
 	}
 }
